backend/requirements.txt
flask-pymongo
flask
flask-cors
python-dotenv